+++ txpipe/__init__.py
@@ -17,5 +17,5 @@ from .photoz_mlz import PZPDFMLZ
 from .covariance import TXFourierGaussianCovariance
 from .metacal_gcr_input import TXMetacalGCRInput
 from .diagnostics import TXDiagnosticPlots
-from .rowe_stats import TXRoweStatistics
+from .psf_plots import TXRoweStatistics
 from .exposure_info import TXExposureInfo
